+++ src/roiextractors/extractors/caiman/caimansegmentationextractor.py
@@ -5,6 +5,7 @@ Classes
 CaimanSegmentationExtractor
     A class for extracting segmentation from CaImAn output.
 """
+
 from pathlib import Path
 
 try:
@@ -57,7 +58,8 @@ class CaimanSegmentationExtractor(SegmentationExtractor):
         self._roi_response_dff = self._trace_extractor_read("F_dff")
         self._roi_response_neuropil = self._trace_extractor_read("C")
         self._roi_response_deconvolved = self._trace_extractor_read("S")
-        self._image_correlation = self._summary_image_read()
+        self._image_correlation = self._correlation_image_read()
+        self._image_mean = self._summary_image_read()
         self._sampling_frequency = self._dataset_file["params"]["data"]["fr"][()]
         self._image_masks = self._image_mask_sparse_read()
 
@@ -111,10 +113,17 @@ class CaimanSegmentationExtractor(SegmentationExtractor):
         if field in self._dataset_file["estimates"]:
             return lazy_ops.DatasetView(self._dataset_file["estimates"][field]).lazy_transpose()
 
-    def _summary_image_read(self):
-        """Read the summary image (Cn) from the estimates dataset of the h5py file."""
+    def _correlation_image_read(self):
+        """Read correlation image Cn."""
         if self._dataset_file["estimates"].get("Cn"):
             return np.array(self._dataset_file["estimates"]["Cn"])
+
+    def _summary_image_read(self):
+        """Read summary image mean."""
+        if self._dataset_file["estimates"].get("b"):
+            FOV_shape = self._dataset_file["params"]["data"]["dims"][()]
+            b_sum = self._dataset_file["estimates"]["b"][:].sum(axis=1)
+            return np.array(b_sum).reshape(FOV_shape, order="F")
 
     def get_accepted_list(self):
         accepted = self._dataset_file["estimates"]["idx_components"]