+++ src/roiextractors/extractors/miniscopeimagingextractor/__init__.py
@@ -10,4 +10,5 @@ Classes
 MiniscopeImagingExtractor
     An ImagingExtractor for the Miniscope video (.avi) format.
 """
+
 from .miniscopeimagingextractor import MiniscopeImagingExtractor